requirements.txt
requests
pyinstaller
pywebview[gtk]; sys_platform == "linux"
pywebview; sys_platform != "linux"
desktop-notifier
pystray
termcolor